Add files via upload
Gold Atlantis , credit providing firm , observed there is rise in number of defaulters. Firm is interested in those customers who are low probability to become a defaulter.
 As an Analyst , we need to analyze the credit data of Gold Atlantis and figure out various  patterns in the data and make interpretation .
The main objective of this analysis is to help the firm to identify the those customers who have lower probability of becoming a defaulter.